+++ vtm/src/org/oscim/theme/XmlThemeBuilder.java
@@ -926,7 +926,7 @@ public class XmlThemeBuilder extends DefaultHandler {
             else if ("map-background".equals(name)) {
                 mapBackground = Color.parseColor(value);
                 if (mThemeCallback != null)
-                    mapBackground = mThemeCallback.getColor(mapBackground);
+                    mapBackground = mThemeCallback.getColor(null, mapBackground);
 
             } else if ("base-stroke-width".equals(name))
                 baseStrokeWidth = Float.parseFloat(value);